refactor: add alert service (DD-3)
