+++ src/app/features/auth/interceptors/auth.interceptor.ts
@@ -1,6 +1,6 @@
 import { HttpInterceptorFn } from '@angular/common/http';
 import { inject } from '@angular/core';
-import { LocalStorageService } from '@core/services/local-storage.service';
+import { LocalStorageService } from '@core/services/local-storage/local-storage.service';
 import { KEY_USER_TOKEN } from '@shared/constants';
 
 export const authInterceptor: HttpInterceptorFn = (req, next) => {
